Printf

+++ aux.c
@@ -43,7 +43,7 @@ int process_int(va_list args)
  * @num: Non-negative integer
  * Return: Number of digits printed
  */
-long print_number(long num)
+int print_number(int num)
 {
 	int count = 0;
 	char digit;
